rename yinglong to yinglong1 (#862)
