+++ examples/yinglong1/predict_12layers.py
@@ -22,8 +22,8 @@ import pandas as pd
 from omegaconf import DictConfig
 from packaging import version
 
-from examples.yinglong.plot import save_plot_weather_from_dict
-from examples.yinglong.predictor import YingLongPredictor
+from examples.yinglong1.plot import save_plot_weather_from_dict
+from examples.yinglong1.predictor import YingLongPredictor
 from ppsci.utils import logger
 
 
@@ -46,23 +46,18 @@ def inference(cfg: DictConfig):
     predictor = YingLongPredictor(cfg)
 
     # load data
-    # HRRR Crop use 69 atmospheric variable，their index in the dataset is from 0 to 68.
-    # The variable name is "z50", "z100",  "z150", "z200", "z250", "z300", "z400", "z500",
-    # "z600", "z700",  "z850", "z925", "z1000", "t50", "t100", "t150", "t200", "t250",
-    # "t300", "t400", "t500", "t600", "t700", "t850", "t925", "t1000", "s50", "s100",
-    # "s150", "s200",  "s250", "s300", "s400", "s500", "s600", "s700", "s850", "s925",
-    # "s1000", "u50", "u100", "u150", "u200", "u250", "u300", "u400", "u500", "u600",
-    # "u700", "u850", "u925", "u1000", "v50", "v100", "v150", "v200", "v250", "v300",
-    # "v400", "v500", "v600", "v700", "v850", "v925", "v1000",  "mslp", "u10", "v10",
-    # "t2m",
+    # HRRR Crop use 24 atmospheric variable，their index in the dataset is from 0 to 23.
+    # The variable name is 'z50', 'z500', 'z850', 'z1000', 't50', 't500', 't850', 'z1000',
+    # 's50', 's500', 's850', 's1000', 'u50', 'u500', 'u850', 'u1000', 'v50', 'v500',
+    # 'v850', 'v1000', 'mslp', 'u10', 'v10', 't2m'.
     input_file = h5py.File(cfg.INFER.input_file, "r")["fields"]
     nwp_file = h5py.File(cfg.INFER.nwp_file, "r")["fields"]
 
-    # input_data.shape: (1, 69, 440, 408)
+    # input_data.shape: (1, 24, 440, 408)
     input_data = input_file[0:1]
-    # nwp_data.shape: # (num_timestamps, 69, 440, 408)
+    # nwp_data.shape: # (num_timestamps, 24, 440, 408)
     nwp_data = nwp_file[0:num_timestamps]
-    # ground_truth.shape: (num_timestamps, 69, 440, 408)
+    # ground_truth.shape: (num_timestamps, 24, 440, 408)
     ground_truth = input_file[1 : num_timestamps + 1]
 
     # create time stamps
@@ -74,7 +69,7 @@ def inference(cfg: DictConfig):
 
     # run predictor
     pred_data = predictor.predict(input_data, time_stamps, nwp_data)
-    pred_data = pred_data.squeeze(axis=1)  # (num_timestamps, 69, 440, 408)
+    pred_data = pred_data.squeeze(axis=1)  # (num_timestamps, 24, 440, 408)
 
     # save predict data
     save_path = osp.join(cfg.output_dir, "result.npy")
@@ -82,7 +77,7 @@ def inference(cfg: DictConfig):
     logger.info(f"Save output to {save_path}")
 
     # plot wind data
-    u10_idx, v10_idx = 66, 67
+    u10_idx, v10_idx = 21, 22
     pred_wind = (pred_data[:, u10_idx] ** 2 + pred_data[:, v10_idx] ** 2) ** 0.5
     ground_truth_wind = (
         ground_truth[:, u10_idx] ** 2 + ground_truth[:, v10_idx] ** 2
@@ -93,7 +88,7 @@ def inference(cfg: DictConfig):
         visu_key = f"Init time: {cfg.INFER.init_time}h\n Ground truth: {i+1}h"
         visu_keys.append(visu_key)
         data_dict[visu_key] = ground_truth_wind[i]
-        visu_key = f"Init time: {cfg.INFER.init_time}h\n YingLong-24 Layers: {i+1}h"
+        visu_key = f"Init time: {cfg.INFER.init_time}h\n YingLong-12 Layers: {i+1}h"
         visu_keys.append(visu_key)
         data_dict[visu_key] = pred_wind[i]
 
@@ -113,7 +108,7 @@ def inference(cfg: DictConfig):
     logger.info(f"Save plot to {cfg.output_dir}")
 
 
-@hydra.main(version_base=None, config_path="./conf", config_name="yinglong_24.yaml")
+@hydra.main(version_base=None, config_path="./conf", config_name="yinglong_12.yaml")
 def main(cfg: DictConfig):
     if cfg.mode == "infer":
         inference(cfg)
